widen grid init to u16 so worlds can exceed 255 per axis
The init loop iterated 0..scene_size as u8, which saturates any
dimension over 255 and silently truncates the world. u16 lifts the
ceiling to 65535; u16 -> f64 is lossless so the Coord conversions are
unchanged.

+++ src/main.rs
@@ -34,8 +34,8 @@ async fn main() {
     let scene = Arc::new(Scene::new(scene_size, scale));
     let dot_factory = DotFactory::new(tx.clone());
 
-    for x in 0..scene_size.x as u8 {
-        for y in 0..scene_size.y as u8 {
+    for x in 0..scene_size.x as u16 {
+        for y in 0..scene_size.y as u16 {
             let pos = Coord {
                 x: x.into(),
                 y: y.into(),
